+++ src/main/java/com/google/cloud/pubsublite/spark/PslWriteDataSourceOptions.java
@@ -32,6 +32,7 @@ import com.google.cloud.pubsublite.internal.wire.PartitionCountWatchingPublisher
 import com.google.cloud.pubsublite.internal.wire.PubsubContext;
 import com.google.cloud.pubsublite.internal.wire.RoutingMetadata;
 import com.google.cloud.pubsublite.internal.wire.SinglePartitionPublisherBuilder;
+import com.google.cloud.pubsublite.spark.internal.PublisherFactory;
 import com.google.cloud.pubsublite.v1.AdminServiceClient;
 import com.google.cloud.pubsublite.v1.AdminServiceSettings;
 import com.google.cloud.pubsublite.v1.PublisherServiceClient;
