+++ src/test/java/com/google/cloud/pubsublite/spark/PslContinuousReaderTest.java
@@ -24,6 +24,9 @@ import com.google.api.core.ApiFutures;
 import com.google.cloud.pubsublite.*;
 import com.google.cloud.pubsublite.internal.CursorClient;
 import com.google.cloud.pubsublite.internal.testing.UnitTestExamples;
+import com.google.cloud.pubsublite.spark.internal.MultiPartitionCommitter;
+import com.google.cloud.pubsublite.spark.internal.PartitionCountReader;
+import com.google.cloud.pubsublite.spark.internal.PartitionSubscriberFactory;
 import com.google.common.collect.ImmutableMap;
 import java.util.Optional;
 import org.junit.Test;
