+++ src/test/java/com/google/cloud/pubsublite/spark/PslMicroBatchReaderTest.java
@@ -28,6 +28,9 @@ import com.google.cloud.pubsublite.Offset;
 import com.google.cloud.pubsublite.Partition;
 import com.google.cloud.pubsublite.internal.CursorClient;
 import com.google.cloud.pubsublite.internal.testing.UnitTestExamples;
+import com.google.cloud.pubsublite.spark.internal.MultiPartitionCommitter;
+import com.google.cloud.pubsublite.spark.internal.PartitionSubscriberFactory;
+import com.google.cloud.pubsublite.spark.internal.PerTopicHeadOffsetReader;
 import com.google.common.collect.ImmutableMap;
 import java.util.Optional;
 import org.junit.Test;
