+++ src/test/java/com/google/cloud/pubsublite/spark/internal/LimitingHeadOffsetReaderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.pubsublite.spark;
+package com.google.cloud.pubsublite.spark.internal;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -30,6 +30,8 @@ import com.google.cloud.pubsublite.Partition;
 import com.google.cloud.pubsublite.internal.TopicStatsClient;
 import com.google.cloud.pubsublite.internal.testing.UnitTestExamples;
 import com.google.cloud.pubsublite.proto.Cursor;
+import com.google.cloud.pubsublite.spark.internal.LimitingHeadOffsetReader;
+import com.google.cloud.pubsublite.spark.internal.PartitionCountReader;
 import com.google.common.testing.FakeTicker;
 import java.util.concurrent.TimeUnit;
 import org.junit.Test;
